~ `Programming Diary`: connect to SQLite with Python
§ The Complete Python🐍/PostgreSQL🐘 Course 2.0
https://pysql.tecladocode.com/

› use `sqlite3` module;
› create one connection;
› use `context manager`.

■ What is a transaction?

A **transaction** is a sequence of operations performed (using one or more SQL statements)
on a database as a `single logical unit of work`.

The effects of all the SQL statements in a **transaction** can be either ALL *committed*
(applied to the database) OR all *rolled back* (undone from the database)!

The easiest way to handle both *committing* and *rolling back* in one go is by using
the **context manager**:
with connection:  # committing and/or rolling back using `context manager`!
    connection.execute("CREATE TABLE entry (content TEXT, date TEXT);")

When you use the **context manager***, `with connection`, it'll automatically *commit*
for you at the end of the *context manager*, or *roll back* if an error was encountered!

» Learning Journal/Programming Diary with Python & SQLite :snake: :minidisc:
› 03-07 How to connect to SQLite with Python | The Complete Python/PostgreSQL Course 2.0
https://pysql.tecladocode.com/section03/lectures/07_connect_sqlite_python/