~ `Programming Diary`: insert data into DB with Python
§ The Complete Python🐍/PostgreSQL🐘 Course 2.0
https://pysql.tecladocode.com/

› We can use the connection to execute a query.
› We can use the `context manager` to automatically commit the changes when we're done.
› We need to be able to tell the query what data we want to insert in the database → `(?, ?)`.

``` sql
connection.execute("INSERT INTO entry VALUES ('This is some test content', '01-01-2020');")
```

We must remember to use the `context manager` so that this insertion is saved into the database:
``` sql
with connection:
    connection.execute("INSERT INTO entry VALUES ('This is some test content', '01-01-2020');")
```

We don't want to always insert the same entry. We do this using the question mark character with `sqlite3`:
``` sql
with connection:
    connection.execute("INSERT INTO entry VALUES (?, ?);", ("This is some test content", "01-01-2020"))
```
The first element of the tuple will be inserted into the query as the first question mark,
and the second one will go in the second question mark.

~ BONUS :books:
» How to check if a table exists in SQLite - Quora
https://www.quora.com/How-can-you-check-if-a-table-exists-in-SQLite

``` sql
-- check if the table with name `entry` exists in SQLite DB
SELECT name FROM sqlite_master WHERE type='table' AND name='entry';
-- This query will return the `name` of the table if it exists in the database.
-- If no rows are returned, then the table does not exist.
```

■ DANGER :warning:

Do not do this: :no_entry_sign:

``` sql
def add_entry(content, date):
    with connection:
        connection.execute(f"INSERT INTO entries VALUES ({content}, {date});")
```
If you do that, you are opening yourself up to a **SQL Injection Attack**.

~ useful links :bookmark: :memo:

database - SQL with SQLite print message command - Stack Overflow
https://stackoverflow.com/questions/35687519/sql-with-sqlite-print-message-command

sql - SQLite syntax for If Else condition - Stack Overflow
https://stackoverflow.com/questions/14580294/sqlite-syntax-for-if-else-condition

How to Check if a Table Exists in SQLite
https://database.guide/how-to-check-if-a-table-exists-in-sqlite/
How to check if a table exists in SQLite - Quora

» Learning Journal/Programming Diary with Python & SQLite :snake: :minidisc:
› 03-10 How to insert data into SQLite with Python | The Complete Python/PostgreSQL Course 2.0
https://pysql.tecladocode.com/section03/lectures/10_python_execute_sqlite_queries/