~ `Programming Diary`: retrieve data with python
§ The Complete Python🐍/PostgreSQL🐘 Course 2.0
https://pysql.tecladocode.com/

› connect, execute, make use of cursor that we get back

When retrieving data we don't need to use the **context manager** because we're not going to commit or rollback anything!
So we'll just use the **connection** to execute the query:

``` sql
cursor = connection.execute("SELECT * FROM entries;")
```

`connection.execute` returns the cursor that it created implicitly to run the query.
It is the **same thing** as doing this:

``` sql
cursor = connection.cursor()
cursor.execute("SELECT * FROM entries;")
```

Now that our cursor has executed the query, it is pointing at the very first row of the results returned by the database.

We just need to use those results. We can:

• Fetch one row;
• Fetch all rows;
• Iterate over the cursor as if it were a list (using a `for` loop).

So with the cursor we can do:

▪ `cursor.fetchone()` - gives you a **tuple** with the contents of the FIRST ROW in the results.
▪ `cursor.fetchall()` - gives you a **list of tuples** with ALL the CONTENTS of the results.
▪ Something like this:

``` sql
for row in cursor:
    print(row)  # tuple of one row
```

› create data retrieval function

› make SQLite give us named results

Whenever SQLite executes a query and builds a result set, it puts each resultant row of data into the cursor as a tuple.
But we can tell it to create an object instead.

``` sql
connection = sqlite3.connect("data.db")
connection.row_factory = sqlite3.Row
```

You can still access the data by index if you prefer, but using `sqlite3.Row` you can also access the data by name.
It's up to you which one to use, but there's no point to using `sqlite3.Row` if you always access data by index.
Just something to remember! :memo:

python - What is the purpose of the row_factory method of an sqlite3 Connection object? - Stack Overflow
https://stackoverflow.com/questions/44009452/what-is-the-purpose-of-the-row-factory-method-of-an-sqlite3-connection-object/44009503#44009503

row_factory -> sqlite3 — DB-API 2.0 interface for SQLite databases — Python 3.12.3 documentation
https://docs.python.org/3/library/sqlite3.html#sqlite3.Connection.row_factory

» Learning Journal/Programming Diary with Python & SQLite :snake: :minidisc:
› 03-12 Retrieve data from SQLite queries with Python | The Complete Python/PostgreSQL Course 2.0
https://pysql.tecladocode.com/section03/lectures/12_python_retrieve_query_results_sqlite/